update: correct implementation of wallet connect
